Make logstash module code consistent with other Elastic stack modules (#8497)
This PR makes the Metricbeat `logstash` module code consistent with other Elastic stack modules by:

* using `FetchContent()` instead of `FetchJSON()` to fetch data from LS APIs before passing it to the event mapping function.
* adding unit tests.

Further consistency around event reporting and error handling will be achieved in #8308.

(cherry picked from commit 909d78995af2610b01f0d5b3e696659d79deedce)
