+++ metricbeat/module/logstash/node/_meta/test/node.700.json
@@ -1,0 +1,41 @@
+{
+    "host": "Shaunaks-MBP-2",
+    "version": "7.0.0-alpha1",
+    "http_address": "127.0.0.1:9600",
+    "id": "7565df20-c3aa-4261-81d5-3b0ab8d15c16",
+    "name": "Shaunaks-MBP-2",
+    "pipelines": {
+        "main": {
+            "workers": 8,
+            "batch_size": 125,
+            "batch_delay": 50,
+            "config_reload_automatic": false,
+            "config_reload_interval": 3000000000,
+            "dead_letter_queue_enabled": false
+        }
+    },
+    "os": {
+        "name": "Mac OS X",
+        "arch": "x86_64",
+        "version": "10.13.6",
+        "available_processors": 8
+    },
+    "jvm": {
+        "pid": 83353,
+        "version": "1.8.0_171",
+        "vm_version": "1.8.0_171",
+        "vm_vendor": "Oracle Corporation",
+        "vm_name": "Java HotSpot(TM) 64-Bit Server VM",
+        "start_time_in_millis": 1538217473740,
+        "mem": {
+            "heap_init_in_bytes": 1073741824,
+            "heap_max_in_bytes": 1037959168,
+            "non_heap_init_in_bytes": 2555904,
+            "non_heap_max_in_bytes": 0
+        },
+        "gc_collectors": [
+            "ParNew",
+            "ConcurrentMarkSweep"
+        ]
+    }
+}
