+++ metricbeat/module/logstash/node_stats/_meta/test/node_stats.650.json
@@ -1,0 +1,129 @@
+{
+    "host": "Shaunaks-MBP-2",
+    "version": "6.5.0",
+    "http_address": "127.0.0.1:9600",
+    "id": "7565df20-c3aa-4261-81d5-3b0ab8d15c16",
+    "name": "Shaunaks-MBP-2",
+    "jvm": {
+        "threads": {
+            "count": 30,
+            "peak_count": 32
+        },
+        "mem": {
+            "heap_used_percent": 34,
+            "heap_committed_in_bytes": 1037959168,
+            "heap_max_in_bytes": 1037959168,
+            "heap_used_in_bytes": 362846688,
+            "non_heap_used_in_bytes": 126413504,
+            "non_heap_committed_in_bytes": 144498688,
+            "pools": {
+                "survivor": {
+                    "peak_used_in_bytes": 35782656,
+                    "used_in_bytes": 23317672,
+                    "peak_max_in_bytes": 35782656,
+                    "max_in_bytes": 35782656,
+                    "committed_in_bytes": 35782656
+                },
+                "old": {
+                    "peak_used_in_bytes": 101436232,
+                    "used_in_bytes": 87297304,
+                    "peak_max_in_bytes": 715849728,
+                    "max_in_bytes": 715849728,
+                    "committed_in_bytes": 715849728
+                },
+                "young": {
+                    "peak_used_in_bytes": 286326784,
+                    "used_in_bytes": 252231712,
+                    "peak_max_in_bytes": 286326784,
+                    "max_in_bytes": 286326784,
+                    "committed_in_bytes": 286326784
+                }
+            }
+        },
+        "gc": {
+            "collectors": {
+                "old": {
+                    "collection_time_in_millis": 147,
+                    "collection_count": 2
+                },
+                "young": {
+                    "collection_time_in_millis": 352,
+                    "collection_count": 8
+                }
+            }
+        },
+        "uptime_in_millis": 118041
+    },
+    "process": {
+        "open_file_descriptors": 95,
+        "peak_open_file_descriptors": 97,
+        "max_file_descriptors": 10240,
+        "mem": {
+            "total_virtual_in_bytes": 7178244096
+        },
+        "cpu": {
+            "total_in_millis": 92452,
+            "percent": 0,
+            "load_average": {
+                "1m": 2.6640625
+            }
+        }
+    },
+    "events": {
+        "in": 0,
+        "filtered": 0,
+        "out": 0,
+        "duration_in_millis": 0,
+        "queue_push_duration_in_millis": 0
+    },
+    "pipelines": {
+        "main": {
+            "events": {
+                "duration_in_millis": 0,
+                "in": 0,
+                "out": 0,
+                "filtered": 0,
+                "queue_push_duration_in_millis": 0
+            },
+            "plugins": {
+                "inputs": [
+                    {
+                        "id": "fc1bd2827edc6b55a3caf4af83c5ac8c5c32b84cb54453bdac8c0d1dcb37ee46",
+                        "events": {
+                            "out": 0,
+                            "queue_push_duration_in_millis": 0
+                        },
+                        "name": "stdin"
+                    }
+                ],
+                "filters": [],
+                "outputs": [
+                    {
+                        "id": "47e3be4432eda99c8d50ea91fdec98aa9c3d5689589ddd3cfdbd62b43746d841",
+                        "events": {
+                            "duration_in_millis": 0,
+                            "in": 0,
+                            "out": 0
+                        },
+                        "name": "stdout"
+                    }
+                ]
+            },
+            "reloads": {
+                "last_error": null,
+                "successes": 0,
+                "last_success_timestamp": null,
+                "last_failure_timestamp": null,
+                "failures": 0
+            },
+            "queue": {
+                "type": "memory"
+            }
+        }
+    },
+    "reloads": {
+        "successes": 0,
+        "failures": 0
+    },
+    "os": {}
+}
